Geodesic+GeodesicExact: Improve accuracy of area calculation (fixing a flaw introduced in version 1.46).  TODO: fix also C, Fortran, Java, JavaScript, Python, MATLAB, and Maxima packages.
The problem introduced in 1.46 was to use the value of omg12 computed in
Lambda12 for computing the area.  At convergence and ignoring round-off,
this is the true value.  However, Lambda12 computes omg12 when solving
the hybrid problem and so the resulting value isn't directly tied to
lam12.  The error for a single area calculation isn't much affected;
but, when computing the area of a polygon, basing the calculation of
omg12 on lam12 allows the errors in the calculations of neighboring
edges to cancel.

Here are the absolute errors in the area calculation for Poland (true
area for WGS84 is 312679715911.98563461373680755101795 m^2).

               series                     exact
      double   long     quad     double   long     quad
1.37 -1.3e-03  2.4e-06  3.4e-14 -6.5e-01 -6.1e-03 -1.0e-17
1.44  3.3e-04  2.5e-06  3.4e-14 -7.6e-01 -5.7e-03 -1.0e-17
1.46 -5.2e-01 -2.0e-03  3.4e-14 -2.5e-01 -2.0e-03 -4.0e-18
1.47  7.0e-04  8.5e-07  3.4e-14 -6.6e-02  5.5e-05 -5.0e-19